Split interface file
